Caustics, bug fixes, more parameters, skybox
