--- Pixel Game/Assets/Scripts/InventoryScreen.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 710b9bb95251a3e419f7dc80ebf170ae
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
