+++ Pixel Game/Assets/Shaders/water/materials/WaterMaterial.mat
@@ -9,7 +9,7 @@ Material:
   m_PrefabAsset: {fileID: 0}
   m_Name: WaterMaterial
   m_Shader: {fileID: 4800000, guid: f486be8620d9af0479b78450c8c33e94, type: 3}
-  m_ShaderKeywords: _ENABLEGROUND_ON _ENABLESKY_ON _SKYREFLECTION_ON
+  m_ShaderKeywords: _ENABLECAUSTICS_ON _ENABLEGROUND_ON _ENABLESKY_ON _SKYREFLECTION_ON
   m_LightmapFlags: 4
   m_EnableInstancingVariants: 0
   m_DoubleSidedGI: 0
@@ -55,6 +55,10 @@ Material:
         m_Texture: {fileID: 0}
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
+    - _Caustics:
+        m_Texture: {fileID: 18700000, guid: 7a2ce15028bcfbc47931a07ac47bfce6, type: 3}
+        m_Scale: {x: 0.43, y: 0.43}
+        m_Offset: {x: 0, y: 0}
     - _Coustics:
         m_Texture: {fileID: 0}
         m_Scale: {x: 1, y: 1}
@@ -77,7 +81,7 @@ Material:
         m_Offset: {x: 0, y: 0}
     - _Ground:
         m_Texture: {fileID: 2800000, guid: c1fe3dc057d13c84ca4ac5ebb0a4ee65, type: 3}
-        m_Scale: {x: 1, y: 1}
+        m_Scale: {x: 0.43, y: 0.43}
         m_Offset: {x: 0, y: 0}
     - _GroundNormalMap:
         m_Texture: {fileID: 0}
@@ -109,6 +113,10 @@ Material:
         m_Offset: {x: 0, y: 0}
     - _Sky:
         m_Texture: {fileID: 2800000, guid: 7f62f2618a037ea49b216ae4d9c0e129, type: 3}
+        m_Scale: {x: 0.35, y: 0.7}
+        m_Offset: {x: 0.2, y: 0.1}
+    - _SkyCube:
+        m_Texture: {fileID: 8900000, guid: 071a1811e584e624db7010cb147a3a63, type: 3}
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     - _SkyTex:
@@ -121,20 +129,26 @@ Material:
         m_Offset: {x: 0, y: 0}
     m_Floats:
     - _BumpScale: 1
-    - _CameraHeight: 0.25
+    - _CameraHeight: 0.5
+    - _CameraHeightScale: 1
+    - _CausticsBrightness: 0.6
+    - _CausticsContrast: 1
     - _Cutoff: 0.5
     - _DetailNormalMapScale: 1
     - _DstBlend: 0
+    - _EnableCaustics: 1
     - _EnableExternalAlpha: 0
     - _EnableGround: 1
     - _EnableSky: 1
     - _FbmGain: 0.3
     - _FbmIterations: -1
     - _FogDensity: 0
+    - _FresnelOffset: 0
+    - _FresnelScale: 1
     - _GlossMapScale: 1
     - _Glossiness: 0.5
     - _GlossyReflections: 1
-    - _GroundParallax: 0.9
+    - _GroundParallax: 0.6
     - _GroundScaleX: 2.3
     - _GroundScaleY: 2.3
     - _LightDistortionFromWaves: 0.05
@@ -158,7 +172,7 @@ Material:
     - _UVSec: 0
     - _WaterOpacity: 0.5
     - _WaterTransparency: 0.5
-    - _WaveAmplitude: 0.5
+    - _WaveAmplitude: 0.3
     - _WaveFbmGain: 0.3
     - _WaveFbmIterations: -1
     - _WaveFrequency: 1
@@ -168,7 +182,7 @@ Material:
     - _blabla: 1
     m_Colors:
     - _Color: {r: 1, g: 1, b: 1, a: 1}
-    - _DirToSun: {r: 1, g: 1, b: 0.6, a: 1}
+    - _DirToSun: {r: 0.2, g: -1, b: 0.6, a: 1}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
     - _Flip: {r: 1, g: 1, b: 1, a: 1}
     - _FogCol: {r: 1, g: 1, b: 1, a: 1}
