+++ Pixel Game/Assets/Shaders/water/textures/skybox.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 24c645945c7185d4f9a15b516970b27d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
