+++ Pixel Game/Assets/Shaders/water/textures/skybox/New Cubemap.cubemap.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a2323a16d40807e47bc0f933d36ab914
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 8900000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
