+++ Pixel Game/UserSettings/EditorUserSettings.asset
@@ -8,6 +8,9 @@ EditorUserSettings:
     RecentlyUsedScenePath-0:
       value: 22424703114646680e0b0227036c731105041d15182d3524121a1c3ef2ec310be1e238eca92f31352d1b
       flags: 0
+    RecentlyUsedScenePath-1:
+      value: 22424703114646680e0b0227036c7219151c1d651f292b20212c2e30e7ee3115acf238e0f323
+      flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
       flags: 0
